Refactor block 6 to display API value and tree count
The changes in this commit refactor the code in block 6 to display the value obtained from an API call. The function `get_api_value()` is added to retrieve the value from the API endpoint `https://tree-nation.com/api/forests/webecode/tree_counter`. The obtained value is then displayed below the image in block 6, along with the label "Arbres plantés". This enhancement improves the functionality and user experience of block 6.